requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
distro==1.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-environ==0.4.5
django-filter==2.4.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-telegrambot==1.0.1
djangorestframework==3.12.2
docker==4.4.4
docker-compose==1.28.5
dockerpty==0.4.1
docopt==0.6.2
geojson==2.5.0
gunicorn==20.0.4
idna==2.10
jsonschema==3.2.0
Markdown==3.3.4
numpy==1.20.2
oauthlib==3.1.0
pandas==1.1.3
paramiko==2.7.2
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyowm==3.2.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.4.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
texttable==1.6.3
tornado==6.1
twilio==6.55.0
tzlocal==2.1
urllib3==1.26.3
websocket-client==0.58.0
